Transforming the project into a Unity's package.
see #1
