--- Assets/Editor/Resources/ShapeEditorGraph.uss
@@ -1,6 +1,0 @@
-﻿GridBackground {
-    --grid-background-color: rgb(32, 32, 32);
-    --line-color: rgba(255,255,255,.1);
-    --thick-line-color: rgba(255,255,255,.3);
-    --spacing: 100;
-}
