+++ Assets/Templates/Square.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 43750f7573e805f4b8e791b1dadd2e96
+guid: 02c939e193bec384ab9deb90c6bc0740
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
