+++ Assets/Templates/Triangle.asset
@@ -9,8 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8bf1bb8ecfe8306439fc3ac26c9f95a7, type: 3}
-  m_Name: Triangle
+  m_Script: {fileID: 11500000, guid: f390f20bf9fcdba46863e197ec4cba41, type: 3}
+  m_Name: Shape Editor Graph
   m_EditorClassIdentifier: 
   points:
   - {x: -1, y: -1}
