+++ Assets/Templates/Triangle.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 71bae1a1a12fc03449bad9a7507a7d0f
+guid: 3724de5e3c5f7bf4281f07422d9af069
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
