+++ Packages/shape-editor/CHANGELOG.md
@@ -1,0 +1,14 @@
+﻿# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [0.1.0] - 2015-10-06
+### Added
+- Unity package to be able to import this project with Unity's package manager.
+
+[Unreleased]: https://github.com/Seabirds-Games/shape-editor/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/Seabirds-Games/shape-editor/releases/tag/v0.1.0
