+++ Packages/shape-editor/CHANGELOG.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 54500594b0be30244b46734e4202d32b
-folderAsset: yes
-DefaultImporter:
+guid: 504c5e7837e25db49af720ddc1277e1a
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
