+++ Packages/shape-editor/Editor.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 54500594b0be30244b46734e4202d32b
+guid: ec399ab083cdc5a4f8b103098c4e779b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
