+++ Packages/shape-editor/Editor/Resources.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 54500594b0be30244b46734e4202d32b
+guid: aff732d38a9da2b488abb267367d0bbe
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
