+++ Packages/shape-editor/Editor/Resources/ShapeEditorGraph.uss.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ddbc8b35aac961147925701b68f71e4a
+guid: 53bca6aa766b42d48b758ebce452157a
 ScriptedImporter:
   internalIDToNameTable: []
   externalObjects: {}
