+++ Packages/shape-editor/Editor/ShapeEditorGraphView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da66d940734146b409fc38163aeb68b4
+guid: ae30ca8d306f0b541828d654d1781108
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
