+++ Packages/shape-editor/Editor/ShapeEditorWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4f9ee04f29ed4fa4ba66b55899cc54d1
+guid: ab737870ba5a05c4a82808a27b7b4cf7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
