+++ Packages/shape-editor/Editor/ShapeManipulator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8bf1bb8ecfe8306439fc3ac26c9f95a7
+guid: 1b510ab8d48868243bffb39a7a21ea23
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
