+++ Packages/shape-editor/Editor/Unity.com.seabirdsgames.shapeeditor.Editor.asmdef
@@ -1,0 +1,4 @@
+﻿{
+  "name": "Shape Custom Editor",
+  "references": ["Shape Runtime"]
+}
