+++ Packages/shape-editor/Editor/Unity.com.seabirdsgames.shapeeditor.Editor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 54500594b0be30244b46734e4202d32b
-folderAsset: yes
-DefaultImporter:
+guid: 21b180ded50e42f4bb5c93bba0dcfc14
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
