+++ Packages/shape-editor/LICENSE.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: b02f3f903fa165946bef2ce4545a5571
-folderAsset: yes
-DefaultImporter:
+guid: d130942c174750140af0012d1d2b711f
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
