+++ Packages/shape-editor/Runtime.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b826abd03ea16774ca6c045b9f91ee0b
+guid: 306d1bf34f2360f42a8d7c765bdfbbfd
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
