+++ Packages/shape-editor/Runtime/ShapeTemplate.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 72655050bced22e4d93e1ef88dae0d76
+guid: f390f20bf9fcdba46863e197ec4cba41
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
