+++ Packages/shape-editor/Runtime/Unity.com.seabirdsgames.shapeeditor.asmdef
@@ -1,0 +1,3 @@
+﻿{
+  "name": "Shape Runtime"
+}
