+++ Packages/shape-editor/Runtime/Unity.com.seabirdsgames.shapeeditor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 54500594b0be30244b46734e4202d32b
-folderAsset: yes
-DefaultImporter:
+guid: 44f73ff20a3c0d54d92799664327ea47
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
