+++ Packages/shape-editor/package.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 54500594b0be30244b46734e4202d32b
-folderAsset: yes
-DefaultImporter:
+guid: 6082a471704021543a1fd7c250e6fd17
+PackageManifestImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
